Refactor tear-off
