+++ packages/freezed/lib/src/templates/abstract_template.dart
@@ -28,59 +28,13 @@ mixin _\$$name$genericsDefinition {
 ${abstractProperties.join()}
 
 $copyWithPrototype
-
 $when
-
 $maybeWhen
-
 $map
-
 $maybeMap
-
 $toJson
 }
-
-class _\$${name}TearOff {
-  const _\$${name}TearOff();
-
-${tearOffs.join()}
-}
-
-const \$$name = _\$${name}TearOff();
 ''';
-  }
-
-  Iterable<String> get tearOffs sync* {
-    for (final targetConstructor in allConstructors) {
-      final ctorName =
-          targetConstructor.isDefault ? 'call' : targetConstructor.name;
-
-      final parameters = StringBuffer();
-      for (final positional
-          in targetConstructor.parameters.allPositionalParameters) {
-        parameters..write(positional.name)..write(',');
-      }
-      for (final named in targetConstructor.parameters.namedParameters) {
-        parameters
-          ..write(named.name)
-          ..write(':')
-          ..write(named.name)
-          ..write(',');
-      }
-
-      var prefix = '';
-      if (targetConstructor.isConst &&
-          genericsParameter.typeParameters.isEmpty &&
-          targetConstructor.parameters.allParameters.isEmpty) {
-        prefix = 'const ';
-      }
-
-      yield '''
-${targetConstructor.redirectedName}$genericsParameter $ctorName$genericsDefinition(${targetConstructor.parameters.asExpanded(showDefaultValue: true)}) {
-  return $prefix ${targetConstructor.redirectedName}$genericsParameter($parameters);
-}
-''';
-    }
   }
 
   String get toJson {
